Refactor server configuration and improve comment readability
This commit involves major changes to server/config/env.go. The main alteration comes from enhancing the way configuration variables are handled. Instead of directly fetching these variables from environment variables, they are now retrieved using the viper library's Get method. This update also refactors the way default values for configuration variables are set. Besides, comments included in the code are made much more detailed and clear, improving the readability and understanding of the code.

+++ server/global/const.go
@@ -115,7 +115,7 @@ const (
 	MaxChannelSize      = 500
 	MaxLuaStatePoolSize = 10000
 
-	MaxActionWorkers = uint(10)
+	MaxActionWorkers = 10
 )
 
 // Log level.
